Converted the feral colony form to an activity
