+++ app/src/main/AndroidManifest.xml
@@ -36,7 +36,10 @@
         <activity
             android:name=".MapsActivity"
             android:label="@string/title_activity_maps"></activity>
-        <activity android:name=".FriendActivity">
+        <activity
+            android:name=".FriendActivity"
+            android:label="Form"></activity>
+        <activity android:name=".LoginActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
