+++ app/src/main/java/com/example/feralfriends/MapsActivity.java
@@ -105,6 +105,8 @@ public class MapsActivity extends FragmentActivity implements OnMapReadyCallback
                             @Override
                             public void onClick(View view)
                             {
+                                Intent intent = FriendActivity.newIntent(getApplicationContext());
+                                startActivity(intent);
                             }
                         });
