--- app/src/main/java/com/example/feralfriends/SingleFragmentActivity.java
@@ -1,39 +1,0 @@
-package com.example.feralfriends;
-
-import android.os.Bundle;
-
-import androidx.annotation.LayoutRes;
-import androidx.appcompat.app.AppCompatActivity;
-import androidx.fragment.app.Fragment;
-import androidx.fragment.app.FragmentManager;
-
-
-public abstract class SingleFragmentActivity extends AppCompatActivity
-{
-    protected abstract Fragment createFragment();
-
-    @LayoutRes
-    protected int getLayoutRes()
-    {
-        return R.layout.activity_fragment;
-    }
-
-    @Override
-    protected void onCreate(Bundle savedInstanceState)
-    {
-        super.onCreate(savedInstanceState);
-        setContentView(getLayoutRes());
-
-        FragmentManager fm = getSupportFragmentManager();
-
-        Fragment fragment = fm.findFragmentById(R.id.fragment_container);
-
-        if(fragment == null)
-        {
-            fragment =  createFragment();
-            fm.beginTransaction()
-                    .add(R.id.fragment_container, fragment)
-                    .commit();
-        }
-    }
-}
