+++ app/src/main/res/layout/activity_friend.xml
@@ -2,78 +2,146 @@
 <LinearLayout
     xmlns:android="http://schemas.android.com/apk/res/android"
     android:orientation="vertical"
-    android:layout_margin="0dp"
+    android:layout_margin="5dp"
     android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:weightSum="10"
-    android:background="#ffffff">
+    android:layout_height="match_parent">
 
-        <ImageView
+    <LinearLayout
+        android:orientation="horizontal"
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:layout_marginTop="16dp"
+        android:layout_marginLeft="0dp"
+        android:layout_marginRight="0dp">
+        <LinearLayout
             android:orientation="vertical"
             android:layout_width="match_parent"
-            android:layout_height="0dp"
-            android:layout_weight="5"
-            android:src = "@drawable/background_logo"/>
+            android:layout_height="match_parent"
+            android:layout_gravity="center"
+            android:padding="0dp">
+
+            <ImageButton
+                android:id="@+id/friend_photo_button"
+                android:layout_width="match_parent"
+                android:layout_height="140dp"
+                android:background="@android:color/darker_gray"
+                android:cropToPadding="true"
+                android:scaleType="center" />
+        </LinearLayout>
+
+    </LinearLayout>
     <LinearLayout
-        android:orientation="vertical"
+        android:orientation="horizontal"
         android:layout_width="match_parent"
-        android:layout_height="0dp"
-        android:layout_weight="2"
-        android:weightSum="4">
+        android:layout_height="wrap_content"
+        android:layout_marginTop="16dp"
+        android:layout_marginLeft="0dp"
+        android:layout_marginRight="0dp">
 
         <EditText
-            android:layout_gravity="center"
+            style="?android:listSeparatorTextViewStyle"
             android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:layout_marginLeft="16dp"
-            android:layout_marginRight="16dp"
-            android:layout_weight="2"
-            android:inputType="textEmailAddress"
-            android:hint="Email/Username" />
-        <EditText
-            android:layout_gravity="center"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:layout_marginLeft="16dp"
-            android:layout_marginRight="16dp"
-            android:layout_weight="2"
-            android:inputType="textPassword"
-            android:hint="Password" />
+            android:layout_height="match_parent"
+            android:fontFamily="cursive"
+            android:gravity="center"
+            android:text="@string/friend_name_label"
+            android:id="@+id/friend_title"
+            android:textSize="24dp" />
+
     </LinearLayout>
+
     <LinearLayout
-        android:orientation="vertical"
+        android:orientation="horizontal"
         android:layout_width="match_parent"
-        android:layout_height="0dp"
-        android:layout_weight="3"
-        android:weightSum="4">
-        <Button
-            android:id="@+id/login_button"
-            android:layout_width="300dp"
-            android:layout_gravity="center"
-            android:layout_height="wrap_content"
-            android:layout_marginTop="16dp"
-            android:layout_marginBottom="5dp"
-            android:layout_marginLeft="16dp"
-            android:layout_marginRight="16dp"
-            android:text="LOGIN"
-            android:background="@color/buttonColor"/>
-        <com.google.android.gms.common.SignInButton
-            android:id="@+id/google_button"
-            android:layout_width="305dp"
-            android:layout_gravity="center"
-            android:layout_height="wrap_content"
-            android:layout_marginLeft="16dp"
-            android:layout_marginRight="16dp"
-            android:text="Text"
-            android:layout_alignParentTop="true"
-            android:layout_margin="10dp"
-            android:textSize="18sp" />
-        <TextView
-            android:id="@+id/text_view"
-            android:layout_gravity="center"
+        android:layout_height="40dp"
+        android:layout_marginTop="16dp"
+        android:layout_marginLeft="0dp"
+        android:layout_marginRight="0dp">
+        <ToggleButton
+            android:id="@+id/button_tnr"
+            android:layout_width="40dp"
+            android:layout_height="40dp"
+            android:background="@drawable/button_favorite"
+            android:checked="false"
+            android:text=""
+            android:textOff=""
+            android:textOn=""
+            android:layout_marginRight="30dp"
+            android:layout_marginLeft="16dp"/>
+        <ImageView
+            android:layout_width="40dp"
+            android:layout_height="match_parent"
+            android:src="@drawable/cat_icon_sleep"
+            android:layout_marginRight="2dp"/>
+
+        <EditText
+            android:id="@+id/number_friends"
+            android:layout_width="40dp"
+            android:layout_height="match_parent"
+            android:fontFamily="cursive"
+            android:gravity="center"
+            android:text="@string/number_of_cats"
+            android:textSize="16dp" />
+
+        <RatingBar
+            style="@style/ImageRatingBar"
             android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:text="forgot Password..?" />
+            android:layout_height="match_parent"
+            android:layout_gravity="center_horizontal"
+            android:layout_marginTop="20dp"
+            android:progressDrawable="@drawable/ic_friendliness_rating"
+            android:numStars="5"
+            android:stepSize="0.2"
+            android:rating="3.0"/>
+
     </LinearLayout>
+    <TextView
+        style="?android:listSeparatorTextViewStyle"
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:text="@string/friend_details_label"
+        />
+
+    <Button
+        android:id="@+id/friend_date"
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:background="#87c0f1"
+        android:fontFamily="cursive"
+        android:text="Feed"
+        android:textSize="24sp"
+        android:textStyle="bold" />
+
+    <Button
+        android:id="@+id/friend_report"
+        android:layout_width="match_parent"
+        android:layout_height="70dp"
+        android:fadingEdge="horizontal|vertical"
+        android:fontFamily="cursive"
+        android:text="@string/friend_report_text"
+        android:textSize="18sp"
+        android:textStyle="bold" />
+
+    <Button
+        android:id="@+id/save_button"
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:background="#87f2b0"
+        android:fadingEdge="horizontal|vertical"
+        android:fontFamily="cursive"
+        android:text="@string/save_friend_text"
+        android:textSize="18sp"
+        android:textStyle="bold" />
+
+    <Button
+        android:id="@+id/delete_button"
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:background="#87f2b0"
+        android:fadingEdge="horizontal|vertical"
+        android:fontFamily="cursive"
+        android:text="@string/delete_friend"
+        android:textSize="18sp"
+        android:textStyle="bold" />
 
 </LinearLayout>
